Updating / debugging functions in transCSSR_bc.
Updating filter_and_pred_probs, run_transCSSR, draw_dot_singlearrows, and save_states:

* filter_and_pred_probs:

Updated this function to allow for filtering of states / predictive probabilities even when the eM / eT becomes un-synchronized. That is, we allow for the case that a forbidden input/output pair is observed, in which case we return to the un-sychronized state and restart the filtering process from that time point.

* run_transCSSR:

Modified how the words of length < L_max - 1 are removed from the causal states after the homogenization step. Before, I removed these words, but without also removing their *counts*. This was not a problem when the counts of words are non-sparse, but led to issues when a certain causal state only had a very small number of words, and all of those words resulted from the shorter histories. This led to some over-counting of the word probabilities.

* draw_dot_singlearrows:

Now we use *all* histories of length either L_max - 1 or L_max to determine transitions out of a causal state. Before, I only used length L_max words when no length L_max - 1 words were in the causal state. However, this missed some of the possible transitions out of a causal state that are induced by the length L_max words.

* save_states:

Same idea as in draw_dot_singlearrows: now instead of using only a *single* history to determine the possible transitions, we use all of the histories in the causal state to determine transitions.